Move function avgratings and getallcomments to reviewscontroller back and change route api for these functions
