+++ front/src/components/services/userAPI.js
@@ -9,16 +9,6 @@ function getProfileAgent(id) {
     .get(USERS_URL + "/profile/agent/" + id)
     .then((response) => response.data);
 }
-function getAvgRatings(id) {
-  return axios
-    .get(USERS_URL + "/profile/agent/" + id + "/avgratings")
-    .then((response) => response.data);
-}
-function getCommentsByAgent(id) {
-  return axios
-    .get(USERS_URL + "/" + id + "/comments")
-    .then((response) => response.data);
-}
 function getDestinationByUser(id, destinationId) {
   return axios
     .get(USERS_URL + "/" + id + "/destination/" + destinationId)
@@ -28,7 +18,5 @@ function getDestinationByUser(id, destinationId) {
 export default {
   getUser,
   getProfileAgent,
-  getAvgRatings,
   getDestinationByUser,
-  getCommentsByAgent,
 };
