+++ front/src/config.js
@@ -10,6 +10,8 @@ export const USERS_URL = API_URL + "/users";
 
 export const MESSAGES_URL = API_URL + "/messages";
 
+export const REVIEWS_URL = API_URL + "/reviews";
+
 export const USER_DESTINATIONS_URL = API_URL + "/destinations";
 
 export const BOOKINGS_URL = API_URL + "/bookings";
